+++ src/libc/stdlib.c
@@ -21,7 +21,7 @@ int count_digits(int64_t num) {
 }
 
 /* itoa: write the digits of "num" into "str". "str" needs to have enough space */
-void itoa(char* str, int num) {
+void itoa(char* str, int64_t num) {
     int sign   = 0;
     int digits = count_digits(num);
 
@@ -61,7 +61,7 @@ int ipow(int b, int e) {
 /* itoan: write the first "max_digits" of "num" (at max) into "str". "str" needs to
  * have enough space. Useful for making sure you won't write out of bounds. Keep in
  * mind that max_digits does not include the null terminator. */
-void itoan(char* str, int num, size_t max_digits) {
+void itoan(char* str, int64_t num, size_t max_digits) {
     if (max_digits <= 0) {
         str[0] = '\0';
         return;
